+++ src/main/java/edu/ksu/canvas/model/Account.java
@@ -13,13 +13,13 @@ import java.io.Serializable;
 public class Account extends BaseCanvasModel implements Serializable {
     public static final long serialVersionUID = 1L;
 
-    private Integer id;
+    private Long id;
     private String name;
-    private Integer parentAccountId; // The account's parent ID, or null if this is the root account
-    private Integer rootAccountId; // The ID of the root account, or null if this is the root account
-    private Integer defaultStorageQuotaMb;
-    private Integer defaultUserStorageQuotaMb;
-    private Integer defaultGroupStorageQuotaMb;
+    private Long parentAccountId; // The account's parent ID, or null if this is the root account
+    private Long rootAccountId; // The ID of the root account, or null if this is the root account
+    private Long defaultStorageQuotaMb;
+    private Long defaultUserStorageQuotaMb;
+    private Long defaultGroupStorageQuotaMb;
     private String defaultTimeZone;
     private String sisAccountId;
     private String integrationId; // currently unused, expect it to be null
@@ -29,11 +29,11 @@ public class Account extends BaseCanvasModel implements Serializable {
     private String workflowState; //todo: Should be an enum
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -46,46 +46,46 @@ public class Account extends BaseCanvasModel implements Serializable {
         this.name = name;
     }
 
-    public Integer getParentAccountId() {
+    public Long getParentAccountId() {
         return parentAccountId;
     }
 
-    public void setParentAccountId(Integer parentAccountId) {
+    public void setParentAccountId(Long parentAccountId) {
         this.parentAccountId = parentAccountId;
     }
 
-    public Integer getRootAccountId() {
+    public Long getRootAccountId() {
         return rootAccountId;
     }
 
-    public void setRootAccountId(Integer rootAccountId) {
+    public void setRootAccountId(Long rootAccountId) {
         this.rootAccountId = rootAccountId;
     }
 
     @CanvasField(postKey = "default_storage_quota_mb")
-    public Integer getDefaultStorageQuotaMb() {
+    public Long getDefaultStorageQuotaMb() {
         return defaultStorageQuotaMb;
     }
 
-    public void setDefaultStorageQuotaMb(Integer defaultStorageQuotaMb) {
+    public void setDefaultStorageQuotaMb(Long defaultStorageQuotaMb) {
         this.defaultStorageQuotaMb = defaultStorageQuotaMb;
     }
 
     @CanvasField(postKey = "default_user_storage_quota_mb")
-    public Integer getDefaultUserStorageQuotaMb() {
+    public Long getDefaultUserStorageQuotaMb() {
         return defaultUserStorageQuotaMb;
     }
 
-    public void setDefaultUserStorageQuotaMb(Integer defaultUserStorageQuotaMb) {
+    public void setDefaultUserStorageQuotaMb(Long defaultUserStorageQuotaMb) {
         this.defaultUserStorageQuotaMb = defaultUserStorageQuotaMb;
     }
 
     @CanvasField(postKey = "default_group_storage_quota_mb")
-    public Integer getDefaultGroupStorageQuotaMb() {
+    public Long getDefaultGroupStorageQuotaMb() {
         return defaultGroupStorageQuotaMb;
     }
 
-    public void setDefaultGroupStorageQuotaMb(Integer defaultGroupStorageQuotaMb) {
+    public void setDefaultGroupStorageQuotaMb(Long defaultGroupStorageQuotaMb) {
         this.defaultGroupStorageQuotaMb = defaultGroupStorageQuotaMb;
     }
