+++ src/main/java/edu/ksu/canvas/model/Enrollment.java
@@ -16,7 +16,7 @@ public class Enrollment extends BaseCanvasModel implements Serializable {
     private static final long serialVersionUID = 3L;
 
     private long id;
-    private Integer courseId;
+    private Long courseId;
     private String sisCourseId;
     private String courseIntegrationId;
     private String courseSectionId;
@@ -25,10 +25,10 @@ public class Enrollment extends BaseCanvasModel implements Serializable {
     private String enrollmentState;
     private Boolean limitPrivilegesToCourseSection;
     private String sisImportId;
-    private Integer rootAccountId;
+    private Long rootAccountId;
     private String type;
     private String userId;
-    private Integer associatedUserId;
+    private Long associatedUserId;
     private String role;
     private Date updatedAt;
     private Date startAt;
@@ -57,11 +57,11 @@ public class Enrollment extends BaseCanvasModel implements Serializable {
         this.roleId = roleId;
     }
 
-    public Integer getCourseId() {
+    public Long getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Integer courseId) {
+    public void setCourseId(Long courseId) {
         this.courseId = courseId;
     }
 
@@ -131,11 +131,11 @@ public class Enrollment extends BaseCanvasModel implements Serializable {
         this.sisImportId = sisImportId;
     }
 
-    public Integer getRootAccountId() {
+    public Long getRootAccountId() {
         return rootAccountId;
     }
 
-    public void setRootAccountId(Integer rootAccountId) {
+    public void setRootAccountId(Long rootAccountId) {
         this.rootAccountId = rootAccountId;
     }
 
@@ -157,11 +157,11 @@ public class Enrollment extends BaseCanvasModel implements Serializable {
         this.userId = userId;
     }
 
-    public Integer getAssociatedUserId() {
+    public Long getAssociatedUserId() {
         return associatedUserId;
     }
 
-    public void setAssociatedUserId(Integer associatedUserId) {
+    public void setAssociatedUserId(Long associatedUserId) {
         this.associatedUserId = associatedUserId;
     }
