+++ src/main/java/edu/ksu/canvas/model/assignment/AssignmentGroup.java
@@ -9,18 +9,18 @@ import edu.ksu.canvas.model.BaseCanvasModel;
 public class AssignmentGroup extends BaseCanvasModel implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    Integer id;
+    Long id;
     String name;
-    Integer position;
+    Long position;
     Double groupWeight;
     String sisSourceId;
     List<Assignment> assignments;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -34,11 +34,11 @@ public class AssignmentGroup extends BaseCanvasModel implements Serializable {
     }
 
     @CanvasField(postKey = "position", array = false)
-    public Integer getPosition() {
+    public Long getPosition() {
         return position;
     }
 
-    public void setPosition(Integer position) {
+    public void setPosition(Long position) {
         this.position = position;
     }
