+++ src/main/java/edu/ksu/canvas/model/assignment/Submission.java
@@ -14,11 +14,11 @@ import java.util.List;
 public class Submission extends BaseCanvasModel implements Serializable {
     private static final long serialVersionUID = 2L;
 
-    private Integer id;
-    private Integer assignmentId;
+    private Long id;
+    private Long assignmentId;
     private Assignment assignment;
     private Course course;
-    private Integer attempt;
+    private Long attempt;
     private String body;
     private String grade;
     private Boolean gradeMatchesCurrentSubmission;
@@ -29,27 +29,27 @@ public class Submission extends BaseCanvasModel implements Serializable {
     private String submissionType;
     private Date submittedAt;
     private String url;
-    private Integer userId;
-    private Integer gradeId;
+    private Long userId;
+    private Long gradeId;
     private User user;
     private Boolean late;
     private Boolean assigmentVisible;
     private Boolean excused;
     private String workflowState;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getAssignmentId() {
+    public Long getAssignmentId() {
         return assignmentId;
     }
 
-    public void setAssignmentId(Integer assignmentId) {
+    public void setAssignmentId(Long assignmentId) {
         this.assignmentId = assignmentId;
     }
 
@@ -70,11 +70,11 @@ public class Submission extends BaseCanvasModel implements Serializable {
     }
 
     @CanvasField(postKey = "attempt", array = false)
-    public Integer getAttempt() {
+    public Long getAttempt() {
         return attempt;
     }
 
-    public void setAttempt(Integer attempt) {
+    public void setAttempt(Long attempt) {
         this.attempt = attempt;
     }
 
@@ -158,19 +158,19 @@ public class Submission extends BaseCanvasModel implements Serializable {
         this.url = url;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getGradeId() {
+    public Long getGradeId() {
         return gradeId;
     }
 
-    public void setGradeId(Integer gradeId) {
+    public void setGradeId(Long gradeId) {
         this.gradeId = gradeId;
     }
